+++ src/index.ts
@@ -2,7 +2,7 @@
  * big-amount: BigInt-based rational number library focused on accounting
  *
  * @license Apache-2.0
- * @copyright 2021 LiosK
+ * @copyright 2021-2022 LiosK
  * @packageDocumentation
  */
 
@@ -463,7 +463,7 @@ export class BigAmount {
    * Adds `others` to `this`. This method is conceptually equivalent to
    * `f.add(others[0]).add(others[1])...`, except for optimization.
    *
-   * @category Arithmetic Operation
+   * @category Optimized Arithmetic Operation
    */
   batchAdd(others: BigAmount[]): BigAmount {
     // take subtotals by denominator and then sum them up
@@ -500,7 +500,7 @@ export class BigAmount {
    * Adds `other` to `this`, keeping the denominator unchanged. This method is
    * equivalent to `f.add(other).quantize(f.den, roundingMode)`.
    *
-   * @category Arithmetic Operation
+   * @category Optimized Arithmetic Operation
    */
   fixedAdd(
     other: BigAmount,
@@ -525,7 +525,7 @@ export class BigAmount {
    * Subtracts `other` from `this`, keeping the denominator unchanged. This
    * method is equivalent to `f.sub(other).quantize(f.den, roundingMode)`.
    *
-   * @category Arithmetic Operation
+   * @category Optimized Arithmetic Operation
    */
   fixedSub(
     other: BigAmount,
@@ -550,7 +550,7 @@ export class BigAmount {
    * Multiplies `this` by `other`, keeping the denominator unchanged. This
    * method is equivalent to `f.mul(other).quantize(f.den, roundingMode)`.
    *
-   * @category Arithmetic Operation
+   * @category Optimized Arithmetic Operation
    */
   fixedMul(
     other: BigAmount,
@@ -559,6 +559,72 @@ export class BigAmount {
     return new BigAmount(
       divInt(this.num * other.num, other.den, roundingMode),
       this.den
+    );
+  }
+
+  /**
+   * Divides `this` by `other`, keeping the denominator unchanged. This method
+   * is equivalent to `f.div(other).quantize(f.den, roundingMode)`.
+   *
+   * @category Optimized Arithmetic Operation
+   */
+  fixedDiv(
+    other: BigAmount,
+    roundingMode: RoundingMode = "HALF_EVEN"
+  ): BigAmount {
+    if (other.num === 0n) {
+      throw new RangeError("denominator is zero");
+    }
+    return new BigAmount(
+      divInt(this.num * other.den, other.num, roundingMode),
+      this.den
+    );
+  }
+
+  /**
+   * Multiplies `this` by `other`, resetting the denominator to `newDen`. This
+   * method is equivalent to `f.mul(other).quantize(newDen, roundingMode)` and
+   * is typically useful to multiply a quantity by unit price to calculate the
+   * dollar amount at a specific precision.
+   *
+   * @category Optimized Arithmetic Operation
+   */
+  quantMul(
+    other: BigAmount,
+    newDen: bigint,
+    roundingMode: RoundingMode = "HALF_EVEN"
+  ): BigAmount {
+    const den = this.den * other.den;
+    return new BigAmount(
+      den === newDen
+        ? this.num * other.num
+        : divInt(this.num * other.num * newDen, den, roundingMode),
+      newDen
+    );
+  }
+
+  /**
+   * Divides `this` by `other`, resetting the denominator to `newDen`. This
+   * method is equivalent to `f.div(other).quantize(newDen, roundingMode)` and
+   * is typically useful to divide a dollar amount by quantity to calculate the
+   * unit price at a specific precision.
+   *
+   * @category Optimized Arithmetic Operation
+   */
+  quantDiv(
+    other: BigAmount,
+    newDen: bigint,
+    roundingMode: RoundingMode = "HALF_EVEN"
+  ): BigAmount {
+    if (other.num === 0n) {
+      throw new RangeError("denominator is zero");
+    }
+    const den = this.den * other.num;
+    return new BigAmount(
+      den === newDen
+        ? this.num * other.den
+        : divInt(this.num * other.den * newDen, den, roundingMode),
+      newDen
     );
   }