+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Maintenance
+
+- Updated dev dependencies.
+
 ## v2.0.1 - 2021-12-13
 
 ### Maintenance